requirements.txt
Django==4.2.11
djangorestframework==3.15.2
dj-database-url==2.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.6
wheel==0.38.1
whitenoise==6.5.0
python-dotenv==1.0.1
